requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
SQLAlchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
python-dotenv>=1.0
pydantic>=2.7
asyncpg>=0.29